requirements.txt
pympi-ling>=1.71
pycantonese>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
spacy>=3.7.0
python-Levenshtein>=0.21.0
tqdm>=4.66.0
seaborn>=0.13.0
transformers==4.57.3
torch==2.9.1
openai==2.14.0
scipy==1.16.3